fixed alignment of listings grid and navbar
