+++ src/App.css
@@ -32,3 +32,16 @@ h1 {
 .hover-cursor {
   cursor: pointer;
 }
+
+/* Listings Page */
+.rental-card-img {
+  width: 100%;
+  height: 225px;
+}
+
+.rental-card {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-left: 13%;
+}
